+++ src/InfrastructureModels.jl
@@ -12,8 +12,10 @@ const LOGGER = getlogger(@__MODULE__)
 
 # Register the module level logger at runtime so that folks can access the logger via `getlogger(InfrastructureModels)`
 # NOTE: If this line is not included then the precompiled `Infrastructure.LOGGER` won't be registered at runtime.
-__init__() = Memento.register(LOGGER)
-
+function __init__()
+    Memento.register(LOGGER)
+    Memento.config(LOGGER, "info")
+end
 
 include("core/data.jl")
 include("core/relaxation_scheme.jl")